Split out Table context
